pyproject.toml
[tool.poetry]
name = "united-states-of-browsers-2"
version = "0.1.0"
description = ""
authors = ["Kshitij Chawla <kc.insight.pi@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^1.9.0"
PyYAML = "^6.0"
Pygments = "^2.12.0"
devtools = "^0.8.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
